# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parali Se Paisa Action Platform API",
    "description": "Offline-first prototype API. Demo records remain marked synthetic; this is not an operational government service.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/public/certificates/{certificate_id}": {
      "get": {
        "tags": [
          "public certificate verification"
        ],
        "summary": "Public Verify",
        "operationId": "public_verify_api_v1_public_certificates__certificate_id__get",
        "parameters": [
          {
            "name": "certificate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Certificate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/certificates/{certificate_id}/file": {
      "get": {
        "tags": [
          "public certificate verification"
        ],
        "summary": "Public Pdf",
        "operationId": "public_pdf_api_v1_public_certificates__certificate_id__file_get",
        "parameters": [
          {
            "name": "certificate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Certificate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system-status": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "System Status",
        "operationId": "system_status_api_v1_system_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Product",
        "operationId": "product_api_v1_product_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSnapshot"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pickup/{field_id}": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Request Pickup",
        "operationId": "request_pickup_api_v1_pickup__field_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/evidence/{scenario}": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Demo Evidence",
        "description": "Replay one historical fixture through the ordinary observation API logic.",
        "operationId": "demo_evidence_api_v1_demo_evidence__scenario__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scenario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductSnapshot": {
        "properties": {
          "fields": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Fields"
          },
          "balers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Balers"
          },
          "buyers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Buyers"
          },
          "jobs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Jobs"
          },
          "runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Runs"
          },
          "verification": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Verification"
          },
          "certificates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Certificates"
          },
          "allocations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Allocations"
          },
          "pickup_requests": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pickup Requests"
          },
          "stats": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Stats"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "fields",
          "balers",
          "buyers",
          "jobs",
          "runs",
          "verification",
          "certificates",
          "allocations",
          "pickup_requests",
          "stats",
          "generated_at"
        ],
        "title": "ProductSnapshot"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}